Increase flexibility of ForageElement hierarchy
This is the third step in the public ecommerce repo that needs to
take place in order to make sharing code between the private pos
repo and this repo pain free; all while allowing for both repos
to evolve independently.

This hierarchy will make it easy to keep ecom things in the
ecom-ForagePinEditText and pos things in the pos-ForagePINEditText
