+++ forage-android/src/main/java/com/joinforage/forage/android/core/services/vault/AbstractVaultSubmitter.kt
@@ -1,7 +1,6 @@
 package com.joinforage.forage.android.core.services.vault
 
 import android.content.Context
-import com.joinforage.forage.android.core.services.launchdarkly.VaultType
 import com.joinforage.forage.android.core.services.telemetry.Log
 import com.joinforage.forage.android.core.services.telemetry.UserAction
 import com.joinforage.forage.android.core.services.telemetry.VaultProxyResponseMonitor
@@ -11,9 +10,8 @@ import com.joinforage.forage.android.core.services.ForageConstants
 import com.joinforage.forage.android.core.services.forageapi.network.ForageApiResponse
 import com.joinforage.forage.android.core.services.forageapi.network.ForageError
 import com.joinforage.forage.android.core.services.forageapi.network.UnknownErrorApiResponse
-import com.joinforage.forage.android.ecom.services.vault.bt.BasisTheoryPinSubmitter
-import com.joinforage.forage.android.ecom.ui.ForagePINEditText
-import com.joinforage.forage.android.ecom.services.vault.vgs.VgsPinSubmitter
+import com.joinforage.forage.android.core.services.VaultType
+import com.joinforage.forage.android.core.ui.element.ForagePinElement
 
 internal val IncompletePinError = ForageApiResponse.Failure.fromError(
     ForageError(400, "user_error", "Invalid EBT Card PIN entered. Please enter your 4-digit PIN.")
@@ -31,16 +29,16 @@ internal open class VaultSubmitterParams(
 
 internal interface VaultSubmitter {
     suspend fun submit(params: VaultSubmitterParams): ForageApiResponse<String>
-
-    fun getVaultType(): VaultType
 }
 
-internal abstract class AbstractVaultSubmitter<VaultResponse>(
+internal abstract class AbstractVaultSubmitter(
     protected val context: Context,
-    protected val foragePinEditText: ForagePINEditText,
+    protected val foragePinEditText: ForagePinElement,
     protected val logger: Log,
-    private val vaultType: VaultType
 ) : VaultSubmitter {
+
+    abstract val vaultType: VaultType
+
     // interface methods
     override suspend fun submit(params: VaultSubmitterParams): ForageApiResponse<String> {
         logger.addAttribute("payment_method_ref", params.paymentMethod.ref)
@@ -99,27 +97,10 @@ internal abstract class AbstractVaultSubmitter<VaultResponse>(
         return forageResponse
     }
 
-    override fun getVaultType(): VaultType {
-        return vaultType
-    }
-
     // abstract methods
     internal abstract fun parseEncryptionKey(encryptionKeys: EncryptionKeys): String
     internal abstract suspend fun submitProxyRequest(vaultProxyRequest: VaultProxyRequest): ForageApiResponse<String>
     internal abstract fun getVaultToken(paymentMethod: PaymentMethod): String?
-
-    /**
-     * @return [UnknownErrorApiResponse] if the response is a vault error, or null if it is not
-     */
-    internal abstract fun toVaultErrorOrNull(vaultResponse: VaultResponse): ForageApiResponse.Failure?
-    internal abstract fun toForageErrorOrNull(vaultResponse: VaultResponse): ForageApiResponse.Failure?
-    internal abstract fun toForageSuccessOrNull(vaultResponse: VaultResponse): ForageApiResponse.Success<String>?
-
-    /**
-     * @return A string containing the raw error details of the vault error.
-     * To be used for internal error reporting.
-     */
-    internal abstract fun parseVaultErrorMessage(vaultResponse: VaultResponse): String
 
     // concrete methods
     protected open fun buildProxyRequest(
@@ -145,36 +126,6 @@ internal abstract class AbstractVaultSubmitter<VaultResponse>(
         return tokensList[index]
     }
 
-    protected fun vaultToForageResponse(vaultResponse: VaultResponse): ForageApiResponse<String> {
-        if (vaultResponse == null) {
-            logger.e("[$vaultType] Received null response from $vaultType")
-            return UnknownErrorApiResponse
-        }
-
-        val vaultError = toVaultErrorOrNull(vaultResponse)
-        if (vaultError != null) {
-            val rawVaultError = parseVaultErrorMessage(vaultResponse)
-            logger.e("[$vaultType] Received error from $vaultType: $rawVaultError")
-            return vaultError
-        }
-
-        val forageApiErrorResponse = toForageErrorOrNull(vaultResponse)
-        if (forageApiErrorResponse != null) {
-            val firstError = forageApiErrorResponse.errors[0]
-            logger.e("[$vaultType] Received ForageError from $vaultType: $firstError")
-            return forageApiErrorResponse
-        }
-
-        val forageApiSuccess = toForageSuccessOrNull(vaultResponse)
-        if (forageApiSuccess != null) {
-            logger.i("[$vaultType] Received successful response from $vaultType")
-            return forageApiSuccess
-        }
-        logger.e("[$vaultType] Received malformed response from $vaultType: $vaultResponse")
-
-        return UnknownErrorApiResponse
-    }
-
     protected fun buildBaseRequestBody(vaultProxyRequest: VaultProxyRequest): HashMap<String, Any> {
         return hashMapOf(
             ForageConstants.RequestBody.CARD_NUMBER_TOKEN to vaultProxyRequest.vaultToken
@@ -182,21 +133,6 @@ internal abstract class AbstractVaultSubmitter<VaultResponse>(
     }
 
     internal companion object {
-        internal fun create(foragePinEditText: ForagePINEditText, logger: Log): VaultSubmitter {
-            val vaultType = foragePinEditText.getVaultType()
-            if (vaultType == VaultType.BT_VAULT_TYPE) {
-                return BasisTheoryPinSubmitter(
-                    context = foragePinEditText.context,
-                    foragePinEditText = foragePinEditText,
-                    logger = logger
-                )
-            }
-            return VgsPinSubmitter(
-                context = foragePinEditText.context,
-                foragePinEditText = foragePinEditText,
-                logger = logger
-            )
-        }
 
         const val TOKEN_DELIMITER = ","
