+++ forage-android/src/main/java/com/joinforage/forage/android/ecom/ui/ForagePINEditText.kt
@@ -4,28 +4,17 @@ import android.app.Application
 import android.content.Context
 import android.graphics.Typeface
 import android.util.AttributeSet
-import android.view.Gravity
-import android.view.ViewGroup
-import android.widget.EditText
-import android.widget.LinearLayout
-import com.basistheory.android.view.TextElement
-import com.joinforage.forage.android.core.ForageConfigNotSetException
 import com.joinforage.forage.android.core.services.launchdarkly.LDManager
 import com.joinforage.forage.android.R
-import com.joinforage.forage.android.core.services.launchdarkly.VaultType
-import com.joinforage.forage.android.core.ui.element.AbstractForageElement
+import com.joinforage.forage.android.core.services.VaultType
 import com.joinforage.forage.android.core.services.EnvConfig
 import com.joinforage.forage.android.core.ui.element.ForageConfig
-import com.joinforage.forage.android.core.ui.element.SimpleElementListener
-import com.joinforage.forage.android.core.ui.element.StatefulElementListener
-import com.joinforage.forage.android.core.ui.element.state.PinElementState
-import com.joinforage.forage.android.core.ui.getLogoImageViewLayout
 import com.joinforage.forage.android.core.services.telemetry.Log
+import com.joinforage.forage.android.core.ui.element.ForagePinElement
 import com.joinforage.forage.android.ecom.ui.vault.bt.BTVaultWrapper
-import com.joinforage.forage.android.ecom.ui.vault.VaultWrapper
+import com.joinforage.forage.android.core.ui.VaultWrapper
 import com.joinforage.forage.android.ecom.ui.vault.vgs.VGSVaultWrapper
 import com.launchdarkly.sdk.android.LDConfig
-import com.verygoodsecurity.vgscollect.widget.VGSEditText
 
 /**
  * A [ForageElement] that securely collects a card PIN. You need a [ForagePINEditText] to call
@@ -65,54 +54,16 @@ class ForagePINEditText @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = R.attr.foragePanEditTextStyle
-) : AbstractForageElement<PinElementState>(context, attrs, defStyleAttr) {
-    private val _linearLayout: LinearLayout
+) : ForagePinElement(context, attrs, defStyleAttr) {
     private val btVaultWrapper: BTVaultWrapper
     private val vgsVaultWrapper: VGSVaultWrapper
 
-    override fun showKeyboard() {
-        if (vault.getVaultType() == VaultType.VGS_VAULT_TYPE) {
-            vault.getVGSEditText().showKeyboard()
-        }
-        if (vault.getVaultType() == VaultType.BT_VAULT_TYPE) {
-            vault.getTextElement().showKeyboard(0)
-        }
-    }
-
-    /**
-     * The `vault` property acts as an abstraction for the actual code
-     * in ForagePINEditText, allowing it to work with a non-nullable
-     * result determined by the choice between BT or VGS. This choice
-     * depends on Launch Darkly and requires knowledge of the environment,
-     * which is determined by `forageConfig` set on this instance.
-     *
-     * The underlying value for `vault` is stored in `_SET_ONLY_vault`.
-     * This backing property is set only after `ForageConfig` has been
-     * initialized for this instance. If `vault` is accessed before
-     * `_SET_ONLY_vault` is set, a runtime exception is thrown.
-     */
-    private var _SET_ONLY_vault: VaultWrapper? = null
-    private val vault: VaultWrapper
-        get() {
-            if (_SET_ONLY_vault == null) {
-                throw ForageConfigNotSetException(
-                    """You are attempting invoke a method a ForageElement before setting
-                    it's ForageConfig. Make sure to call
-                    myForageElement.setForageConfig(forageConfig: ForageConfig) 
-                    immediately on your ForageElement before you call any other methods.
-                    """.trimIndent()
-                )
-            }
-            return _SET_ONLY_vault!!
-        }
+    override fun showKeyboard() = vault.showKeyboard()
 
     init {
         context.obtainStyledAttributes(attrs, R.styleable.ForagePINEditText, defStyleAttr, 0)
             .apply {
                 try {
-                    setWillNotDraw(false)
-                    orientation = VERTICAL
-                    gravity = Gravity.CENTER
 
                     // at this point in time, we do not know the environment and
                     // we are operating and thus do not know whether to add
@@ -126,27 +77,13 @@ class ForagePINEditText @JvmOverloads constructor(
                     // ensure all wrappers init with the
                     // same typeface (or the attributes)
                     btVaultWrapper.typeface = vgsVaultWrapper.typeface
-
-                    val elementWidth: Int = getDimensionPixelSize(R.styleable.ForagePINEditText_elementWidth, ViewGroup.LayoutParams.MATCH_PARENT)
-                    val elementHeight: Int = getDimensionPixelSize(R.styleable.ForagePINEditText_elementHeight, ViewGroup.LayoutParams.WRAP_CONTENT)
-
-                    _linearLayout = LinearLayout(context)
-                    _linearLayout.layoutParams = ViewGroup.LayoutParams(elementWidth, elementHeight)
-                    _linearLayout.orientation = VERTICAL
-                    _linearLayout.gravity = Gravity.CENTER
                 } finally {
                     recycle()
                 }
             }
     }
 
-    override fun initWithForageConfig(forageConfig: ForageConfig) {
-        // Must initialize DD at the beginning of each render function. DD requires the context,
-        // so we need to wait until a context is present to run initialization code. However,
-        // we have logging all over the SDK that relies on the render happening first.
-        val logger = Log.getInstance()
-        logger.initializeDD(context, forageConfig)
-
+    override fun determineBackingVault(forageConfig: ForageConfig, logger: Log) : VaultWrapper {
         // initialize Launch Darkly singleton
         val ldMobileKey = EnvConfig.fromForageConfig(forageConfig).ldMobileKey
         val ldConfig = LDConfig.Builder().mobileKey(ldMobileKey).build()
@@ -154,65 +91,11 @@ class ForagePINEditText @JvmOverloads constructor(
 
         // decide on a vault provider and the corresponding vault wrapper
         val vaultType = LDManager.getVaultProvider(logger)
-        _SET_ONLY_vault = if (vaultType == VaultType.BT_VAULT_TYPE) {
+        return if (vaultType == VaultType.BT_VAULT_TYPE) {
             btVaultWrapper
         } else {
             vgsVaultWrapper
         }
-
-        _linearLayout.addView(vault.getUnderlying())
-        _linearLayout.addView(getLogoImageViewLayout(context))
-        addView(_linearLayout)
-
-        logger.i("[View] ForagePINEditText successfully rendered")
-    }
-
-    override fun clearText() {
-        vault.clearText()
-    }
-
-    // While the events that ForageElements expose mirrors the
-    // blur, focus, change etc events of an Android view,
-    // they represent different abstractions. Our users need to
-    // interact with the ForageElement abstraction and not the
-    // implementation details of which Android view we use.
-    // Therefore we expose novel set listener methods instead of
-    // overriding the convention setOn*Listener
-    override fun setOnFocusEventListener(l: SimpleElementListener) {
-        vault.setOnFocusEventListener(l)
-    }
-    override fun setOnBlurEventListener(l: SimpleElementListener) {
-        vault.setOnBlurEventListener(l)
-    }
-    override fun setOnChangeEventListener(l: StatefulElementListener<PinElementState>) {
-        vault.setOnChangeEventListener(l)
-    }
-
-    override fun getElementState(): PinElementState {
-        return vault.manager.getState()
-    }
-
-    internal fun getVaultType(): VaultType {
-        return vault.getVaultType()
-    }
-
-    internal fun getTextInputEditText(): VGSEditText {
-        return vault.getVGSEditText()
-    }
-
-    internal fun getTextElement(): TextElement {
-        return vault.getTextElement()
-    }
-
-    internal fun getForageTextElement(): EditText {
-        return vault.getForageTextElement()
-    }
-
-    override fun setTextColor(textColor: Int) {
-        vault.setTextColor(textColor)
-    }
-    override fun setTextSize(textSize: Float) {
-        vault.setTextSize(textSize)
     }
 
     override var typeface: Typeface?
@@ -223,19 +106,4 @@ class ForagePINEditText @JvmOverloads constructor(
             btVaultWrapper.typeface = value
             vgsVaultWrapper.typeface = value
         }
-    override fun setHint(hint: String) {
-        vault.setHint(hint)
-    }
-    override fun setHintTextColor(hintTextColor: Int) {
-        vault.setHintTextColor(hintTextColor)
-    }
-    override fun setBoxStrokeColor(boxStrokeColor: Int) {
-        // no-ops for now
-    }
-    override fun setBoxStrokeWidth(boxStrokeWidth: Int) {
-        // no-ops for now
-    }
-    override fun setBoxStrokeWidthFocused(boxStrokeWidth: Int) {
-        // no-ops for now
-    }
 }
