+++ forage-android/src/main/java/com/joinforage/forage/android/ecom/ui/vault/vgs/VGSVaultWrapper.kt
@@ -8,17 +8,19 @@ import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.Gravity
 import android.view.ViewGroup
-import android.widget.EditText
 import android.widget.LinearLayout
-import com.basistheory.android.view.TextElement
-import com.joinforage.forage.android.core.services.launchdarkly.VaultType
+import com.joinforage.forage.android.core.services.VaultType
 import com.joinforage.forage.android.core.ui.element.state.PinElementStateManager
 import com.joinforage.forage.android.core.services.ForageConstants
+import com.joinforage.forage.android.core.services.vault.AbstractVaultSubmitter
+import com.joinforage.forage.android.core.services.telemetry.Log
 import com.joinforage.forage.android.core.ui.getBoxCornerRadiusBottomEnd
 import com.joinforage.forage.android.core.ui.getBoxCornerRadiusBottomStart
 import com.joinforage.forage.android.core.ui.getBoxCornerRadiusTopEnd
 import com.joinforage.forage.android.core.ui.getBoxCornerRadiusTopStart
-import com.joinforage.forage.android.ecom.ui.vault.VaultWrapper
+import com.joinforage.forage.android.ecom.services.vault.vgs.VgsPinSubmitter
+import com.joinforage.forage.android.core.ui.VaultWrapper
+import com.joinforage.forage.android.core.ui.element.ForagePinElement
 import com.verygoodsecurity.vgscollect.core.model.state.FieldState
 import com.verygoodsecurity.vgscollect.core.storage.OnFieldStateChangeListener
 import com.verygoodsecurity.vgscollect.view.card.validation.rules.VGSInfoRule
@@ -31,6 +33,7 @@ internal class VGSVaultWrapper @JvmOverloads constructor(
 ) : VaultWrapper(context, attrs, defStyleAttr) {
     private var _internalEditText: VGSEditText
     override val manager: PinElementStateManager = PinElementStateManager.forEmptyInput()
+    override val vaultType: VaultType = VaultType.VGS_VAULT_TYPE
 
     init {
         context.obtainStyledAttributes(attrs, com.joinforage.forage.android.R.styleable.ForagePINEditText, defStyleAttr, 0)
@@ -122,28 +125,23 @@ internal class VGSVaultWrapper @JvmOverloads constructor(
             }
     }
 
-    override fun getVaultType(): VaultType {
-        return VaultType.VGS_VAULT_TYPE
-    }
-
     override fun clearText() {
         _internalEditText.setText("")
     }
 
-    override fun getVGSEditText(): VGSEditText {
-        return _internalEditText
-    }
+    override fun getTextElement(): VGSEditText = _internalEditText
 
-    override fun getTextElement(): TextElement {
-        throw RuntimeException("Unimplemented for this vault!")
-    }
+    override fun getVaultSubmitter(
+        foragePinElement: ForagePinElement,
+        logger: Log
+    ): AbstractVaultSubmitter = VgsPinSubmitter(
+        foragePinElement.context,
+        foragePinElement,
+        logger
+    )
 
-    override fun getForageTextElement(): EditText {
-        throw RuntimeException("Unimplemented for this vault!")
-    }
-
-    override fun getUnderlying(): VGSEditText {
-        return _internalEditText
+    override fun showKeyboard() {
+        _internalEditText.showKeyboard()
     }
 
     override var typeface: Typeface?
